# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "AppController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/products": {
      "get": {
        "operationId": "ProductController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de productos obtenida exitosamente"
          }
        },
        "summary": "Obtener todos los productos",
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Crear nuevo producto",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        },
        "summary": "Obtener producto por ID",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto eliminado exitosamente"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        },
        "summary": "Eliminar producto (soft delete)",
        "tags": [
          "products"
        ]
      }
    },
    "/products/stats/deleted-percentage": {
      "get": {
        "operationId": "ProductController_getDeletedPercentage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Porcentaje calculado exitosamente"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener porcentaje de productos eliminados",
        "tags": [
          "products"
        ]
      }
    },
    "/products/stats/non-deleted-percentage": {
      "get": {
        "operationId": "ProductController_getNonDeletedPercentage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Porcentaje calculado exitosamente"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener porcentaje de productos no eliminados",
        "tags": [
          "products"
        ]
      }
    },
    "/products/low-stock": {
      "get": {
        "operationId": "ProductController_getLowStockProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de productos con stock bajo"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener productos con stock bajo",
        "tags": [
          "products"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso"
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        },
        "summary": "Iniciar sesión",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          }
        },
        "summary": "Registrar usuario",
        "tags": [
          "auth"
        ]
      }
    },
    "/external/products": {
      "get": {
        "operationId": "ExternalApiController_getProductsFromExternal",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Productos obtenidos exitosamente"
          },
          "500": {
            "description": "Error al obtener productos"
          }
        },
        "summary": "Obtener productos desde API externa (Contentful)",
        "tags": [
          "external-api"
        ]
      }
    }
  },
  "info": {
    "title": "Contentful NestJS API",
    "description": "API para gestión de productos con integración Contentful",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "auth",
      "description": "Autenticación de usuarios"
    },
    {
      "name": "products",
      "description": "Gestión de productos"
    },
    {
      "name": "external-api",
      "description": "API externa Contentful"
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "description": "SKU único del producto",
            "example": "PROD-001"
          },
          "name": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "iPhone 15 Pro"
          },
          "brand": {
            "type": "string",
            "description": "Marca del producto",
            "example": "Apple"
          },
          "model": {
            "type": "string",
            "description": "Modelo del producto",
            "example": "iPhone 15 Pro Max"
          },
          "category": {
            "type": "string",
            "description": "Categoría del producto",
            "example": "Smartphones"
          },
          "color": {
            "type": "string",
            "description": "Color del producto",
            "example": "Natural Titanium"
          },
          "price": {
            "type": "number",
            "description": "Precio del producto",
            "example": 999.99
          },
          "currency": {
            "type": "string",
            "description": "Moneda del precio",
            "example": "USD"
          },
          "stock": {
            "type": "number",
            "description": "Cantidad en stock",
            "example": 50
          },
          "externalId": {
            "type": "string",
            "description": "ID externo del producto (para sincronización)",
            "example": "contentful-123"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Indica si el producto está eliminado",
            "example": false
          }
        },
        "required": [
          "sku",
          "name",
          "brand",
          "model",
          "category",
          "color",
          "price"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email del usuario",
            "example": "usuario@ejemplo.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario",
            "example": "password123",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email del usuario",
            "example": "usuario@ejemplo.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario",
            "example": "password123",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      }
    }
  }
}